first working draft of analysis component of voxelwise atlas connectivity
